feat: Enhance large disk image support and migrate installer to Inno Setup
Refactors AdfImage to use `long` for block addressing, removing the 4 GiB total image size limit and introducing sparse in-memory images for efficient handling of very large HDFs.

Improves OFS/FFS filesystem handling by introducing streaming data access for files, preventing large files from being loaded entirely into memory. This enables opening files larger than available RAM within mounted partitions.

Enforces the AmigaDOS 4 GiB partition size limit for OFS/FFS filesystems during HDF creation.

Replaces the WiX Toolset-based installer with Inno Setup, simplifying the build process and reducing external dependencies.

+++ src/AdfXplorer.Core/FileSystems/Ofs/OfsBlocks.cs
@@ -92,7 +92,7 @@ internal static class AmigaDosRootBlock
         rootBlockNumber = BlockReader.ReadInt32(image.ReadBlock(0), OfsBlockOffsets.Root_BootBlockRootPointer);
         if (rootBlockNumber == 0)
         {
-            rootBlockNumber = image.SectorCount / 2;
+            rootBlockNumber = (int)(image.SectorCount / 2);
         }
 
         if (rootBlockNumber <= 0 || rootBlockNumber >= image.SectorCount)
